requirements.txt
Django
djangorestframework
psycopg2-binary
requests
pandas
spacy
celery
redis
selenium
geopy
